Moved Add Trip functionality to App.js, out of Trip.js - Trip.js now just for Edit Trip
